refactor: spotMapStore (#19)
