+++ src/store/spotMapStore.ts
@@ -1,6 +1,6 @@
 import { create } from "zustand";
 
-interface MapStore {
+interface SpotMapStore {
   activePlaceId: number | null;
   setActivePlaceId: (id: number | null) => void;
   center: { lat: number; lng: number };
@@ -16,7 +16,7 @@ interface SidebarStore {
 /**
  * 지도의 중심점과 선택한 마커의 정보를 관리하는 store
  */
-export const useMapStore = create<MapStore>((set) => ({
+export const useMapStore = create<SpotMapStore>((set) => ({
   activePlaceId: null,
   setActivePlaceId: (id) => set({ activePlaceId: id }),
   center: { lat: 37.5665, lng: 126.978 },
